requirements.txt
python-telegram-bot
appwrite
jikanpy